% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastCooccur-package.R
\docType{package}
\name{FastCooccur-package}
\alias{FastCooccur}
\alias{FastCooccur-package}
\title{FastCooccur: Fast and Scalable exact species cooccurences}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
