% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adjmat.R
\name{generate_adjmat}
\alias{generate_adjmat}
\title{Compute Adjacency Matrix from Cooccur object}
\usage{
generate_adjmat(cooccur, p.adj = "none", verbose = TRUE)
}
\arguments{
\item{cooccur}{A cooccur object retruned by fast_cooccur}

\item{p.adj}{a correction method for the p.value}

\item{verbose}{TRUE or FALSE. If TRUE, display meaningful messages about the progression}
}
\value{
A signed adjacency matrix
}
\description{
Compute Adjacency Matrix from Cooccur object
}
