% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_probaC}
\alias{get_probaC}
\title{Compute cooccurences probabilities over slices}
\usage{
get_probaC(df, mat)
}
\arguments{
\item{df}{a DataFrame of species pairs}

\item{mat}{a species matrix}
}
\description{
Compute cooccurences probabilities over slices
}
\details{
Internal function. Not to be used by itself
}
