% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitycheck.R
\name{sanity_check}
\alias{sanity_check}
\title{Check that the input matrix meet conditions to compute cooccurences}
\usage{
sanity_check(spp_site_mat, strict = FALSE)
}
\arguments{
\item{spp_site_mat}{The species matrix (apecies (rows) x sites (column))}

\item{strict}{TRUE or FALSE. If TRUE, it will throw an error at the first encountered problem and stop code execution.
If FALSE, if will perform all the tests and throw warnings.}
}
\value{
A list of warnings.
}
\description{
Check that the input matrix meet conditions to compute cooccurences
}
