% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_cooccur.R
\name{fast_cooccur}
\alias{fast_cooccur}
\title{Fast Cooccur - compute exact cooccurence betwenn species pairs}
\usage{
fast_cooccur(mat, chunks = 2)
}
\arguments{
\item{mat}{A apecies (rows) x sites (column) matrix.}

\item{chunks}{The number of chunks into which the data should be split.}
}
\value{
A cooccur object
}
\description{
Fast Cooccur - compute exact cooccurence betwenn species pairs
}
